Provide kickstart file for VMware ESXi via CDROM
The e1000 driver is deprecated in VMware ESXi 6.7 and no longer available
in ESXi 7.0. Packer/QEMU does not emulate any other driver type which
VMware ESXi supports without forcing the user to manually create a bridge.

Packer can create an ISO which contains the kickstart file for VMware ESXi
so networking is no longer required to install VMware. The kickstart file
has been renamed to KS.CFG as VMware ESXi requires kickstart files coming
from a CDROM drive to be in all caps with no dashes. The net_device has
been switched to vmxnet3 as a network device is required for installation
however it is no used.