Updated machine sizing, removed extra whitespace (#76)
I updated the machine sizing for the ubuntu flat and lvm templates, as I
found I was unable to reliably build images with 1024MB of RAM and 4GB
of disk. I found I needed to bump both to 2048 MB of RAM and 8GB of disk
to get the machines to reliably build and not crash with a subiquity
error. I also removed extra white space in the autoinstall templates
that may trip up subiquity.