install-custom-packages/chroot: derive Apt proxy from installer (#331)
Custom packages are processed inside `chroot` after unpacking the image.
If the network requires a proxy _(either MaaS-hosted or external)_, the
transaction will fail _[without these changes or transparent proxying]_.

The `chroot` in the new install _does not_ benefit from the `cloud-init`
configuration of _Apt_ in the network-booted _(live)_ environment. These
changes take advantage of the `/run` mount _(and `jq`)_ to carry this
information into the newly-created installation.