refactor: pin qemu plugin version (#332)
Starting with version 1.1.2, the qemu plugin no longer waits for the
machine to shutdown if there's no `communicator` set in the template.
Most packer-maas templates rely on this behavior.

Let's pin the plugin to the last working version until we find out
whether this is a regression in the plug-in or an undocumented behavior
that we were inadvertently exploiting.

Fixes #323