Add ubuntu cloudimg customization example for BlueField (#338)
It is possible to use the ubuntu-cloudimg template to build images that
can run on BlueField DPUs.
This commit adds a customization script that installs all NVIDIA
software (bluefield kernel, modules, tools, and runtime software, based
on the [bfb-build Dockerfile for DOCA 2.9
LTS](https://github.com/Mellanox/bfb-build/blob/lts-2.9.x/ubuntu/22.04/Dockerfile))
required to deploy a ready to use Ubuntu image on these devices with
MAAS.
The Readme and Makefile have been updated to describe how this
customization script can be used. It's important to note that BlueField
images must be created with `ubuntu_series=jammy` (at least for DOCA 2.9
LTS) and `architecture=arm64`.
The customization script includes the netplan configuration for the
tmfifo_net0 interface and the definition of OVS briges for the host
interfaces representors, because MAAS cannot control them as of today.

Signed-off-by: Davide Trentin <davide.trentin@canonical.com>